Modifed the SVG plugin normaliser to generate the checksums of the exported files for the Xena Metadata section.
As SVGs are open and aren't base64 encoded there was a problem where
the inital SVG and exported SVG dont have the same checksum, this was
thanks to things like the different places Xena puts line endings. 
To combat this there is now a new help method in the
AbstractNormaliser called exportAndGenerateChecksum() which will
export the yet to be finished Xena file we are generating to get the
checksum.
NOTE: This should only be used as a last resort as it isn't a neat
solution.