Created a XenaListener interface so classes can register themselves to receive warnings. These classes can be registered through the Xena class/API.
As to not add a public method in the Xena class to fire the events, as
this is supposed to be internal rather than external, a
XenaEventHandler class was added to handle these events. This class
has to live in the xena.core package so it has access to the firing
events in the Xena class. This class it to work as a remote control
for events used by internal Xena classes. 

The ExportResult class is now a XenaListener and adds any warnings it
gets to a list.